library model
