+++ contracts/app_attest_user_map/src/state.rs
@@ -1,5 +1,5 @@
 use cosmwasm_schema::cw_serde;
-use cosmwasm_std::Addr;
+use cosmwasm_std::{Addr, Binary};
 use cw_storage_plus::{Item, Map};
 
 #[cw_serde]
@@ -10,14 +10,13 @@ pub enum UserColor {
     White,
 }
 
-#[cw_serde]
-pub struct UserStatus {
-    pub number: u8,
-    pub color: UserColor,
-}
+// #[cw_serde]
+// pub struct UserStatus {
+//     pub number: u8,
+//     pub color: UserColor,
+// }
 
-pub const USER_MAP: Map<Addr, UserStatus> = Map::new("user_map");
+pub const USER_MAP: Map<Addr, Binary> = Map::new("user_map");
 
 
-pub const APP_ATTEST_VERIFICATION_ADDR: Item<Addr> = Item::new("app_attest_verification_addr");
 pub const APP_ID: Item<String> = Item::new("app_id");
